+++ src/main/java/io/fabric8/maven/docker/assembly/DockerAssemblyManager.java
@@ -205,7 +205,7 @@ public class DockerAssemblyManager {
     void verifyGivenDockerfile(File dockerFile, BuildImageConfiguration buildConfig, FixedStringSearchInterpolator interpolator, Logger log) throws IOException {
         AssemblyConfiguration assemblyConfig = buildConfig.getAssemblyConfiguration();
         if (assemblyConfig == null) {
-            return;
+           return;
         }
 
         String name = assemblyConfig.getName();